git-svn-id: svn://srv-dc/svn2/Citrus@176986 20bf7d6a-bd21-9d4d-a3c4-a2c14bb8c68a

+++ Lime/Source/Audio/AudioChannel.cs
@@ -34,7 +34,7 @@ namespace Lime
 		public const int NumBuffers = 8;
 
 		public AudioChannelGroup Group;
-		public int Priority;
+		public float Priority;
 		public DateTime StartupTime;
 		public int Id;
